requirements.txt
pygame
PyOpenGL
PyOpenGL_accelerate
Pillow